Adds a key instruction for using with Swift and Cocoapods 0.36+
Cocoapods 0.36 allows pods written in Swift, but only if the user includes a directive (use_frameworks!) that coverts all the pods into frameworks. Frameworks have to be imported explicitly in Swift and can't simply be including in bridging headers. This commit adds a line that explains that fact, so that newbies like me don't waste hours figuring it out like I did.